+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add brightness detection and display comparison on GUI
 - Add diskcache for resized images to speed up GUI
 - Add background thread and queue to handle image caching for previous/next images
+- Add ability to clear diskcache
 
 ### Changed
 
@@ -31,7 +32,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set maximum parallel workers to number of CPU cores
 - Reuse single ThreadPoolExecutor for most tasks to reduce overhead of recreating threads
 - Switch to diskcache to store similarity metadata
+- Replace dlib face detection with OpenCV DNN implementation, removing dependency
+- Increase page cache to 8 pages and pre-load next 2 pages instead of just one
+- Create face temp directory name using hash instead of filename
+
+### Removed
+
 - Remove mmap cache of open image files to avoid parallel thread access issues
+- Remove blur, brightness and contrast detection at startup
 
 ### Fixed
 
@@ -40,3 +48,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use image file creation date if EXIF date is missing
 - Rescan images if similarity info is missing
 - Fixed fullscreen on MacOS
+- Generate face images with same file extension as source image
+- Remove lru_cache for similar images since simfilter amount can change